+++ ext/opentelemetry-ext-prometheus/src/opentelemetry/ext/prometheus/__init__.py
@@ -79,7 +79,7 @@ from prometheus_client.core import (
     UnknownMetricFamily,
 )
 
-from opentelemetry.metrics import Counter, Measure, Metric
+from opentelemetry.metrics import Counter, Metric, ValueRecorder
 from opentelemetry.sdk.metrics.export import (
     MetricRecord,
     MetricsExporter,
@@ -152,22 +152,22 @@ class CustomCollector:
         metric_name = ""
         if self._prefix != "":
             metric_name = self._prefix + "_"
-        metric_name += self._sanitize(metric_record.metric.name)
+        metric_name += self._sanitize(metric_record.instrument.name)
 
-        if isinstance(metric_record.metric, Counter):
+        if isinstance(metric_record.instrument, Counter):
             prometheus_metric = CounterMetricFamily(
                 name=metric_name,
-                documentation=metric_record.metric.description,
+                documentation=metric_record.instrument.description,
                 labels=label_keys,
             )
             prometheus_metric.add_metric(
                 labels=label_values, value=metric_record.aggregator.checkpoint
             )
         # TODO: Add support for histograms when supported in OT
-        elif isinstance(metric_record.metric, Measure):
+        elif isinstance(metric_record.instrument, ValueRecorder):
             prometheus_metric = UnknownMetricFamily(
                 name=metric_name,
-                documentation=metric_record.metric.description,
+                documentation=metric_record.instrument.description,
                 labels=label_keys,
             )
             prometheus_metric.add_metric(
@@ -176,7 +176,7 @@ class CustomCollector:
 
         else:
             logger.warning(
-                "Unsupported metric type. %s", type(metric_record.metric)
+                "Unsupported metric type. %s", type(metric_record.instrument)
             )
         return prometheus_metric