+++ opentelemetry-api/src/opentelemetry/configuration/__init__.py
@@ -12,9 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# FIXME find a better way to avoid all those "Expression has type "Any"" errors
-# type: ignore
-
 """
 Simple configuration manager
 
@@ -78,6 +75,11 @@ To use the meter provider above, then the
 OpenTelemetry API provided providers are the default ones used if no
 configuration is found in the environment variables).
 
+Configuration values that are exactly ``"True"`` or ``"False"`` will be
+converted to its boolean values of ``True`` and ``False`` respectively.
+
+Configuration values that can be casted to integers or floats will be casted.
+
 This object can be used by any OpenTelemetry component, native or external.
 For that reason, the ``Configuration`` object is designed to be immutable.
 If a component would change the value of one of the ``Configuration`` object
@@ -90,17 +92,23 @@ to override this value instead of changing it.
 
 from os import environ
 from re import fullmatch
+from typing import ClassVar, Dict, Optional, TypeVar, Union
+
+ConfigValue = Union[str, bool, int, float]
+_T = TypeVar("_T", ConfigValue, Optional[ConfigValue])
 
 
 class Configuration:
-    _instance = None
-
-    __slots__ = []
+    _instance = None  # type: ClassVar[Optional[Configuration]]
+    _config_map = {}  # type: ClassVar[Dict[str, ConfigValue]]
 
     def __new__(cls) -> "Configuration":
-        if Configuration._instance is None:
+        if cls._instance is not None:
+            instance = cls._instance
+        else:
 
-            for key, value in environ.items():
+            instance = super().__new__(cls)
+            for key, value_str in environ.items():
 
                 match = fullmatch(
                     r"OPENTELEMETRY_PYTHON_([A-Za-z_][\w_]*)", key
@@ -109,31 +117,47 @@ class Configuration:
                 if match is not None:
 
                     key = match.group(1)
+                    value = value_str  # type: ConfigValue
 
-                    setattr(Configuration, "_{}".format(key), value)
-                    setattr(
-                        Configuration,
-                        key,
-                        property(
-                            fget=lambda cls, key=key: getattr(
-                                cls, "_{}".format(key)
-                            )
-                        ),
-                    )
+                    if value_str == "True":
+                        value = True
+                    elif value_str == "False":
+                        value = False
+                    else:
+                        try:
+                            value = int(value_str)
+                        except ValueError:
+                            pass
+                        try:
+                            value = float(value_str)
+                        except ValueError:
+                            pass
 
-                Configuration.__slots__.append(key)
+                    instance._config_map[key] = value
 
-            Configuration.__slots__ = tuple(Configuration.__slots__)
+            cls._instance = instance
 
-            Configuration._instance = object.__new__(cls)
+        return instance
 
-        return cls._instance
+    def __getattr__(self, name: str) -> Optional[ConfigValue]:
+        return self._config_map.get(name)
 
-    def __getattr__(self, name):
-        return None
+    def __setattr__(self, key: str, val: ConfigValue) -> None:
+        if key == "_config_map":
+            super().__setattr__(key, val)
+        else:
+            raise AttributeError(key)
+
+    def get(self, name: str, default: _T) -> _T:
+        """Use this typed method for dynamic access instead of `getattr`
+
+        :rtype: str or bool or int or float or None
+        """
+        val = self._config_map.get(name, default)
+        return val
 
     @classmethod
-    def _reset(cls):
+    def _reset(cls) -> None:
         """
         This method "resets" the global configuration attributes
 
@@ -141,10 +165,6 @@ class Configuration:
         only.
         """
 
-        for slot in cls.__slots__:
-            if slot in cls.__dict__.keys():
-                delattr(cls, slot)
-                delattr(cls, "_{}".format(slot))
-
-        cls.__slots__ = []
-        cls._instance = None
+        if cls._instance:
+            cls._instance._config_map.clear()  # pylint: disable=protected-access
+            cls._instance = None