+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/__init__.py
@@ -97,9 +97,9 @@ class BoundCounter(metrics_api.BoundCounter, BaseBoundInstrument):
             self.update(value)
 
 
-class BoundMeasure(metrics_api.BoundMeasure, BaseBoundInstrument):
+class BoundValueRecorder(metrics_api.BoundValueRecorder, BaseBoundInstrument):
     def record(self, value: metrics_api.ValueT) -> None:
-        """See `opentelemetry.metrics.BoundMeasure.record`."""
+        """See `opentelemetry.metrics.BoundValueRecorder.record`."""
         if self._validate_update(value):
             self.update(value)
 
@@ -174,15 +174,15 @@ class Counter(Metric, metrics_api.Counter):
     UPDATE_FUNCTION = add
 
 
-class Measure(Metric, metrics_api.Measure):
-    """See `opentelemetry.metrics.Measure`."""
+class ValueRecorder(Metric, metrics_api.ValueRecorder):
+    """See `opentelemetry.metrics.ValueRecorder`."""
 
-    BOUND_INSTR_TYPE = BoundMeasure
+    BOUND_INSTR_TYPE = BoundValueRecorder
 
     def record(
         self, value: metrics_api.ValueT, labels: Dict[str, str]
     ) -> None:
-        """See `opentelemetry.metrics.Measure.record`."""
+        """See `opentelemetry.metrics.ValueRecorder.record`."""
         bound_intrument = self.bind(labels)
         bound_intrument.record(value)
         bound_intrument.release()
@@ -190,8 +190,8 @@ class Measure(Metric, metrics_api.Measure):
     UPDATE_FUNCTION = record
 
 
-class Observer(metrics_api.Observer):
-    """See `opentelemetry.metrics.Observer`."""
+class ValueObserver(metrics_api.ValueObserver):
+    """See `opentelemetry.metrics.ValueObserver`."""
 
     def __init__(
         self,
@@ -257,11 +257,11 @@ class Record:
 
     def __init__(
         self,
-        metric: metrics_api.MetricT,
+        instrument: metrics_api.InstrumentT,
         labels: Dict[str, str],
         aggregator: Aggregator,
     ):
-        self.metric = metric
+        self.instrument = instrument
         self.labels = labels
         self.aggregator = aggregator
 
@@ -270,22 +270,21 @@ class Meter(metrics_api.Meter):
     """See `opentelemetry.metrics.Meter`.
 
     Args:
+        source: The `MeterProvider` that created this meter.
         instrumentation_info: The `InstrumentationInfo` for this meter.
-        stateful: Indicates whether the meter is stateful.
     """
 
     def __init__(
         self,
+        source: "MeterProvider",
         instrumentation_info: "InstrumentationInfo",
-        stateful: bool,
-        resource: Resource = Resource.create_empty(),
     ):
         self.instrumentation_info = instrumentation_info
+        self.batcher = UngroupedBatcher(source.stateful)
+        self.resource = source.resource
         self.metrics = set()
         self.observers = set()
-        self.batcher = UngroupedBatcher(stateful)
         self.observers_lock = threading.Lock()
-        self.resource = resource
 
     def collect(self) -> None:
         """Collects all the metrics created with this `Meter` for export.
@@ -375,10 +374,11 @@ class Meter(metrics_api.Meter):
         description: str,
         unit: str,
         value_type: Type[metrics_api.ValueT],
+        observer_type=Type[metrics_api.ObserverT],
         label_keys: Sequence[str] = (),
         enabled: bool = True,
     ) -> metrics_api.Observer:
-        ob = Observer(
+        ob = observer_type(
             callback,
             name,
             description,
@@ -392,27 +392,35 @@ class Meter(metrics_api.Meter):
             self.observers.add(ob)
         return ob
 
-    def unregister_observer(self, observer: "Observer") -> None:
+    def unregister_observer(self, observer: metrics_api.Observer) -> None:
         with self.observers_lock:
             self.observers.remove(observer)
 
 
 class MeterProvider(metrics_api.MeterProvider):
-    def __init__(self, resource: Resource = Resource.create_empty()):
+    """See `opentelemetry.metrics.MeterProvider`.
+
+    Args:
+        stateful: Indicates whether meters created are going to be stateful
+        resource: Resource for this MeterProvider
+    """
+
+    def __init__(
+        self, stateful=True, resource: Resource = Resource.create_empty(),
+    ):
+        self.stateful = stateful
         self.resource = resource
 
     def get_meter(
         self,
         instrumenting_module_name: str,
-        stateful=True,
         instrumenting_library_version: str = "",
     ) -> "metrics_api.Meter":
         if not instrumenting_module_name:  # Reject empty strings too.
             raise ValueError("get_meter called with missing module name.")
         return Meter(
+            self,
             InstrumentationInfo(
                 instrumenting_module_name, instrumenting_library_version
             ),
-            stateful=stateful,
-            resource=self.resource,
         )