+++ opentelemetry-sdk/src/opentelemetry/sdk/trace/__init__.py
@@ -252,11 +252,6 @@ class Span(trace_api.Span):
             this `Span`.
     """
 
-    # Initialize these lazily assuming most spans won't have them.
-    _empty_attributes = BoundedDict(MAX_NUM_ATTRIBUTES)
-    _empty_events = BoundedList(MAX_NUM_EVENTS)
-    _empty_links = BoundedList(MAX_NUM_LINKS)
-
     def __init__(
         self,
         name: str,
@@ -264,7 +259,7 @@ class Span(trace_api.Span):
         parent: Optional[trace_api.SpanContext] = None,
         sampler: Optional[sampling.Sampler] = None,
         trace_config: None = None,  # TODO
-        resource: None = None,
+        resource: Resource = Resource.create_empty(),
         attributes: types.Attributes = None,  # TODO
         events: Sequence[Event] = None,  # TODO
         links: Sequence[trace_api.Link] = (),
@@ -289,22 +284,20 @@ class Span(trace_api.Span):
 
         self._filter_attribute_values(attributes)
         if not attributes:
-            self.attributes = Span._empty_attributes
+            self.attributes = self._new_attributes()
         else:
             self.attributes = BoundedDict.from_map(
                 MAX_NUM_ATTRIBUTES, attributes
             )
 
-        if events is None:
-            self.events = Span._empty_events
-        else:
-            self.events = BoundedList(MAX_NUM_EVENTS)
+        self.events = self._new_events()
+        if events:
             for event in events:
                 self._filter_attribute_values(event.attributes)
                 self.events.append(event)
 
         if links is None:
-            self.links = Span._empty_links
+            self.links = self._new_links()
         else:
             self.links = BoundedList.from_seq(MAX_NUM_LINKS, links)
 
@@ -324,6 +317,18 @@ class Span(trace_api.Span):
         return '{}(name="{}", context={})'.format(
             type(self).__name__, self.name, self.context
         )
+
+    @staticmethod
+    def _new_attributes():
+        return BoundedDict(MAX_NUM_ATTRIBUTES)
+
+    @staticmethod
+    def _new_events():
+        return BoundedList(MAX_NUM_EVENTS)
+
+    @staticmethod
+    def _new_links():
+        return BoundedList(MAX_NUM_LINKS)
 
     @staticmethod
     def _format_context(context):
@@ -360,7 +365,7 @@ class Span(trace_api.Span):
             f_links.append(f_link)
         return f_links
 
-    def to_json(self):
+    def to_json(self, indent=4):
         parent_id = None
         if self.parent is not None:
             if isinstance(self.parent, Span):
@@ -397,7 +402,7 @@ class Span(trace_api.Span):
         f_span["events"] = self._format_events(self.events)
         f_span["links"] = self._format_links(self.links)
 
-        return json.dumps(f_span, indent=4)
+        return json.dumps(f_span, indent=indent)
 
     def get_context(self):
         return self.context
@@ -407,9 +412,6 @@ class Span(trace_api.Span):
             if not self.is_recording_events():
                 return
             has_ended = self.end_time is not None
-            if not has_ended:
-                if self.attributes is Span._empty_attributes:
-                    self.attributes = BoundedDict(MAX_NUM_ATTRIBUTES)
         if has_ended:
             logger.warning("Setting attribute on ended span.")
             return
@@ -422,6 +424,12 @@ class Span(trace_api.Span):
             # Freeze mutable sequences defensively
             if isinstance(value, MutableSequence):
                 value = tuple(value)
+            if isinstance(value, bytes):
+                try:
+                    value = value.decode()
+                except ValueError:
+                    logger.warning("Byte attribute could not be decoded.")
+                    return
             with self._lock:
                 self.attributes[key] = value
 
@@ -442,9 +450,7 @@ class Span(trace_api.Span):
             if not self.is_recording_events():
                 return
             has_ended = self.end_time is not None
-            if not has_ended:
-                if self.events is Span._empty_events:
-                    self.events = BoundedList(MAX_NUM_EVENTS)
+
         if has_ended:
             logger.warning("Calling add_event() on an ended span.")
             return
@@ -458,7 +464,7 @@ class Span(trace_api.Span):
     ) -> None:
         self._filter_attribute_values(attributes)
         if not attributes:
-            attributes = Span._empty_attributes
+            attributes = self._new_attributes()
         self._add_event(
             Event(
                 name=name,