+++ scripts/coverage.sh
@@ -3,13 +3,25 @@
 set -e
 
 function cov {
-    pytest \
-        --ignore-glob=*/setup.py \
-        --cov ${1} \
-        --cov-append \
-        --cov-branch \
-        --cov-report='' \
-        ${1}
+    if [ ${TOX_ENV_NAME:0:4} == "py34" ]
+    then
+        pytest \
+            --ignore-glob=*/setup.py \
+            --ignore-glob=ext/opentelemetry-ext-opentracing-shim/tests/testbed/* \
+            --cov ${1} \
+            --cov-append \
+            --cov-branch \
+            --cov-report='' \
+            ${1}
+    else
+        pytest \
+            --ignore-glob=*/setup.py \
+            --cov ${1} \
+            --cov-append \
+            --cov-branch \
+            --cov-report='' \
+            ${1}
+    fi
 }
 
 PYTHON_VERSION=$(python -c 'import sys; print(".".join(map(str, sys.version_info[:3])))')
@@ -24,6 +36,7 @@ cov ext/opentelemetry-ext-flask
 cov ext/opentelemetry-ext-requests
 cov ext/opentelemetry-ext-jaeger
 cov ext/opentelemetry-ext-opentracing-shim
+cov ext/opentelemetry-exporter-cloud-trace
 cov ext/opentelemetry-ext-wsgi
 cov ext/opentelemetry-ext-zipkin
 cov docs/examples/opentelemetry-example-app
@@ -31,6 +44,9 @@ cov docs/examples/opentelemetry-example-app
 # aiohttp is only supported on Python 3.5+.
 if [ ${PYTHON_VERSION_INFO[1]} -gt 4 ]; then
     cov ext/opentelemetry-ext-aiohttp-client
+# ext-asgi is only supported on Python 3.5+.
+if [ ${PYTHON_VERSION_INFO[1]} -gt 4 ]; then
+    cov ext/opentelemetry-ext-asgi
 fi
 
 coverage report --show-missing