+++ ext/opentelemetry-ext-zipkin/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 0.8b0
+
+Released 2020-05-27
+
+- Transform resource to tags when exporting
+  ([#707](https://github.com/open-telemetry/opentelemetry-python/pull/707))
+
 ## 0.7b1
 
 Released 2020-05-12