+++ opentelemetry-api/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## Unreleased
 
+- Move stateful from Meter to MeterProvider
+  ([#751](https://github.com/open-telemetry/opentelemetry-python/pull/751))
+- Rename Measure to ValueRecorder in metrics
+  ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))
+- Rename Observer to ValueObserver
+  ([#764](https://github.com/open-telemetry/opentelemetry-python/pull/764))
+
+## 0.8b0
+
+Released 2020-05-27
+
+- Handle boolean, integer and float values in Configuration
+  ([#662](https://github.com/open-telemetry/opentelemetry-python/pull/662))
+- bugfix: ensure status is always string
+  ([#640](https://github.com/open-telemetry/opentelemetry-python/pull/640))
+
 ## 0.7b1
 
 Released 2020-05-12